files_from_working_server/waterscan-api/requirements.txt
gunicorn==19.8.1
aniso8601==3.0.0
astroid==1.6.2
click==6.7
colorama==0.3.9
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.12
PyJWT==1.4.0
pylint==1.8.3
python-dateutil==2.7.3
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.5
virtualenv==15.2.0
Werkzeug==0.14.1
wrapt==1.10.11